Refactor ChangeTheSubject so it is available at index time
Also, ensure it takes as input an array of strings, which will make it
easy to integrate into the indexing process.
