+++ marc_to_solr/lib/traject_config.rb
@@ -10,6 +10,7 @@ require_relative './electronic_portfolio_builder'
 require_relative './alma_reader'
 require_relative './solr_deleter'
 require_relative './access_facet_builder'
+require_relative './change_the_subject'
 require 'stringex'
 require 'library_stdnums'
 require 'time'
