+++ spec/marc_to_solr/lib/change_the_subject_spec.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+##
+# When our catalog records contain outdated subject headings, we need the ability
+# to update them at index time to preferred terms.
+RSpec.describe ChangeTheSubject do
+  context "a replaced term" do
+    let(:subject_term) { "Illegal Aliens" }
+
+    it "suggests a replacement" do
+      expect(described_class.check_for_replacement(subject_term)).to eq "Undocumented Immigrants"
+    end
+  end
+
+  context "a term that has not been replaced" do
+    let(:subject_term) { "Daffodils" }
+
+    it "returns the term unchanged" do
+      expect(described_class.check_for_replacement(subject_term)).to eq subject_term
+    end
+  end
+
+  context "an array of subject terms" do
+    let(:subject_terms) { ["Illegal Aliens", "Workplace Safety"] }
+    let(:fixed_subject_terms) { ["Undocumented Immigrants", "Workplace Safety"] }
+
+    it "changes only the subject terms that have been configured" do
+      expect(described_class.fix(subject_terms)).to eq fixed_subject_terms
+    end
+  end
+end
